refactor: extract partition_storage into its own translation unit
Move partition_storage class from raw_disk_io.cpp into
partition_storage.hpp and partition_storage.cpp. Update
raw_disk_io.hpp to include the new header (replacing the forward
declaration), and add partition_storage.cpp to CMakeLists.txt.
